+++ CHANGELOG.md
@@ -4,3 +4,4 @@ Backbone.BindTo
 Version 1.1.0 (unreleased)
 
 * bindTo can accept raw function
+* bindTo can observe even DOM/jQuery elements